Fixes OPENSTACK-735
The configobj requirement, even though added to requirements.txt breaks
installation with devstack cause the dependency has to be present in
global-requirements.txt.

Because it seems unlikely configobj will just be added to upstream devstack.
And to fix jenkins creating baseimages the script is (Temporaly?) moved
to the upgrade-scripts project.

Closes-Bug: OPENSTACK-735

Change-Id: Ifbf48c2f751b4a0cff43290249ee52790ae196a1