Added script for uninstall nuage folder cr8ed by setup.py install
For the icehouse release we provide a tar.gz along with a setup.py
script which the user runs to install the nuage specfiv icehouse
folder under the python path using setup.py install. Now, during
migration to kilo, the admins will do a apt-get upgrade of the
neutron server. This results in a dpkg warning for the nuage
specific folder since the package manager is unaware of the this
nuage folder created by setup.py install. This script will basically
do the opposite of setup.py install. That is it will find the path
where neutron is installed, find the nuage path and delete the nuage
folder. Run this script before doing an apt-get upgrade of the
neutron.

Change-Id: Ie8405f3e6e8fc10b554a15b8a1d783ef096c109b
Closes-Bug: Openstack-930