Script to support setting override flag on user/group resource
Customer trying to use VSD's LDAP feature along with Openstack in
advance mode, will need to run this script prior to turning on that
feature and after upgrading their VSD to 3.2R5

This script should work with following combination of upgrade:
1. Icehouse + 3.0Rx/3.2R(1-4) -> Icehouse + 3.2R5
2. Icehouse + 3.0Rx/3.2R(1-4) -> Juno + 3.2R5
3. Icehouse + 3.0Rx/3.2R(1-4) -> Kilo + 3.2R5
4. Juno + 3.0x/3.2R(1-4) -> Juno + 3.2R5
5. Juno + 3.0x/3.2R(1-4) -> Kilo + 3.2R5
6. Kilo + 3.2R(1-4) -> Kilo + 3.2R5

Script at this point can be run independantly. It can be incorporated
into a umbreall upgrade script as well; as and when it happens.

Both user and group's 'managementMode' attribute are expected to be changed
from default 'ANY' to 'CMS' once the script is run.

This operation only makes sense for the advanced subnet. But it will be performed
for normal subnet as well.

The way it works is that we walk through all the subnets that we have registered
in VSD (L2domain and Subnet). In the database table subnet_l2dom_mapping we keep
track of nuage user and group id for them. We do a PUT operation on all of them.
First User, than Group.

One thing that is not captured and will not work:
When router is created in openstack, l3domain and zone(s) are created and corresponding
tenant is mapped to user/group in the private zone in VSD. Via this script they are not
going to get updated unless they have subnet attached to it.

Change-Id: I9a8cc107c86af9bc3d35d0d3cb8fe6171a2f553a