Fix for OPENSTACK-1457
Neutron's database configuration is set in such a way that when
no configuration is present, it will default to using an in-memory
database. This is giving users a confusing message saying they are
connected to the database, while in fact it's not the database
they expect. This is fixed by changing the default to something
incorrect, so database setup will fail and give a proper message.

Also removed the --interactive setting of the script as it had
little added value.

Change-Id: I35f75f9f5f19fd3682c4e20ac13cfbac0c2f64e8
Closes-Bug: OPENSTACK-1457