Remove dependency on plugin VSD client, validation of port cms_id ...
and port not found error reporting fix. Currently the upgrade script
has indirect dependency on our plugin VSD client which was causing
the script to hang. Also it assumes that there is only one CMS
using a single subnet, however it need not be the case.
Furthermore, resolved the port not found error logging
with sufficient details.

Change-Id: Ibb357d0d1611b84b8dbd43c7074749056615af2f
Closes-Bug: OPENSTACK-2610