Fix dhcp port creation for dhcp enabled v4 subnets
For multiple v4 subnets in one network, all dhcp ports are deleted
if deleting one of the subnets. They will be recovered by 6.0 upgrade.
This patchset makes the dhcp ports recovery only for internal subnets;
And the dhcp ip is the first availale ip for os managed subnets.
For vsd managed subnets, use the gateway ip from vsd, if it is used by
another port, raise a warning.

Change-Id: I334dbeb9145caf7fe4634d2efdbfa4ae961a92c8
Closes-Bug: OPENSTACK-2742, OPENSTACK-2743